# helloczcz/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "hello world",
        "version": "1.0.0",
        "description": "The Data Set API (DSAPI) allows the public users to discover and search USPTO exported data sets. This is a generic API that allows USPTO users to make any CSV based data files searchable through API. With the help of GET call, it returns the list of data fields that are searchable. With the help of POST call, data can be fetched based on the filters on the field names. Please note that POST call is used to search the actual data. The reason for the POST call is that it allows users to specify any complex search criteria without worry about the GET size limitations as well as encoding of the input parameters.",
        "contact": {
            "name": "Open Data Portal",
            "url": "https://developer.uspto.gov",
            "email": "developer@uspto.gov"
        }
    },
    "servers": [
        {
            "url": "{scheme}://developer.uspto.gov/ds-api",
            "variables": {
                "scheme": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "https",
                    "description": "The Data Set API is accessible via https and http"
                }
            }
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "metadata"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dataSetList"
                                },
                                "example": {
                                    "total": 2,
                                    "apis": [
                                        {
                                            "apiKey": "oa_citations",
                                            "apiVersionNumber": "v1",
                                            "apiUrl": "https://developer.uspto.gov/ds-api/oa_citations/v1/fields",
                                            "apiDocumentationUrl": "https://developer.uspto.gov/ds-api-docs/index.html?url=https://developer.uspto.gov/ds-api/swagger/docs/oa_citations.json"
                                        },
                                        {
                                            "apiKey": "cancer_moonshot",
                                            "apiVersionNumber": "v1",
                                            "apiUrl": "https://developer.uspto.gov/ds-api/cancer_moonshot/v1/fields",
                                            "apiDocumentationUrl": "https://developer.uspto.gov/ds-api-docs/index.html?url=https://developer.uspto.gov/ds-api/swagger/docs/cancer_moonshot.json"
                                        }
                                    ]
                                }
                            }
                        },
                        "description": "Returns a list of data sets"
                    }
                },
                "operationId": "list-data-sets",
                "summary": "List available data sets"
            }
        },
        "/{dataset}/{version}/fields": {
            "get": {
                "tags": [
                    "metadata"
                ],
                "parameters": [
                    {
                        "example": "oa_citations",
                        "name": "dataset",
                        "description": "Name of the dataset.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "required": true
                    },
                    {
                        "example": "v1",
                        "name": "version",
                        "description": "Version of the dataset.",
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "The dataset API for the given version is found and it is accessible to consume."
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "The combination of dataset name and version is not found in the system or it is not published yet to be consumed by public."
                    }
                },
                "operationId": "list-searchable-fields",
                "summary": "Provides the general information about the API and the list of fields that can be used to query the dataset.",
                "description": "This GET API returns the list of all the searchable field names that are in the oa_citations. Please see the 'fields' attribute which returns an array of field names. Each field or a combination of fields can be searched using the syntax options shown below."
            }
        },
        "/{dataset}/{version}/records": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "criteria"
                                ],
                                "type": "object",
                                "properties": {
                                    "criteria": {
                                        "description": "Uses Lucene Query Syntax in the format of propertyName:value, propertyName:[num1 TO num2] and date range format: propertyName:[yyyyMMdd TO yyyyMMdd]. In the response please see the 'docs' element which has the list of record objects. Each record structure would consist of all the fields and their corresponding values.",
                                        "default": "*:*",
                                        "type": "string"
                                    },
                                    "start": {
                                        "description": "Starting record number. Default value is 0.",
                                        "default": 0,
                                        "type": "integer"
                                    },
                                    "rows": {
                                        "description": "Specify number of rows to be returned. If you run the search with default values, in the response you will see 'numFound' attribute which will tell the number of records available in the dataset.",
                                        "default": 100,
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "search"
                ],
                "parameters": [
                    {
                        "name": "version",
                        "description": "Version of the dataset.",
                        "schema": {
                            "default": "v1",
                            "type": "string"
                        },
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "dataset",
                        "description": "Name of the dataset. In this case, the default value is oa_citations",
                        "schema": {
                            "default": "oa_citations",
                            "type": "string"
                        },
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "additionalProperties": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "No matching record found for the given criteria."
                    }
                },
                "operationId": "perform-search",
                "summary": "Provides search capability for the data set with the given search criteria.",
                "description": "This API is based on Solr/Lucense Search. The data is indexed using SOLR. This GET API returns the list of all the searchable field names that are in the Solr Index. Please see the 'fields' attribute which returns an array of field names. Each field or a combination of fields can be searched using the Solr/Lucene Syntax. Please refer https://lucene.apache.org/core/3_6_2/queryparsersyntax.html#Overview for the query syntax. List of field names that are searchable can be determined using above GET api."
            }
        },
        "/abcs": {
            "summary": "Path used to manage the list of abcs.",
            "description": "The REST endpoint/path used to list and create zero or more `abc` entities.  This path contains a `GET` and `POST` operation to perform the list and create tasks, respectively.",
            "servers": [
                {
                    "url": "http://domain.example.com:80/api",
                    "description": "xxx"
                }
            ],
            "get": {
                "requestBody": {
                    "description": "CZXCZX",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            },
                            "examples": {
                                "CZXCXZ": {
                                    "value": "\"some text\""
                                }
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "XxxzX",
                    "DxzXz",
                    "metadata"
                ],
                "parameters": [
                    {
                        "examples": {
                            "xczxczxZXZ": {
                                "value": [
                                    {

                                    },
                                    {

                                    }
                                ]
                            }
                        },
                        "name": "xxx",
                        "description": "zxxxx",
                        "schema": {
                            "type": "array",
                            "items": {

                            }
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "headers": {
                            "XCZXC": {

                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/abc"
                                    }
                                },
                                "examples": {
                                    "XCZXC": {
                                        "value": [
                                            {
                                                "xzx": "zxX"
                                            },
                                            {
                                                "xzx": "fbcxbc"
                                            }
                                        ]
                                    }
                                }
                            },
                            "text/xml": {

                            }
                        },
                        "description": "Successful response - returns an array of `abc` entities."
                    }
                },
                "operationId": "getabcs",
                "summary": "List All abcs",
                "description": "Gets a list of all `abc` entities."
            },
            "post": {
                "requestBody": {
                    "description": "A new `abc` to be created.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/abc"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful response."
                    }
                },
                "operationId": "createabc",
                "summary": "Create a abc",
                "description": "Creates a new instance of a `abc`."
            },
            "parameters": [
                {
                    "examples": {
                        "xczxczxZXZ": {
                            "value": [
                                {

                                },
                                {

                                }
                            ]
                        }
                    },
                    "name": "xxx",
                    "description": "zxxxx",
                    "schema": {
                        "type": "array",
                        "items": {

                        }
                    },
                    "in": "query",
                    "required": true
                },
                {
                    "examples": {
                        "XZXX": {
                            "value": "\"some text\""
                        }
                    },
                    "name": "SCASCSAS",
                    "description": "",
                    "schema": {
                        "type": "string"
                    },
                    "in": "header",
                    "required": true
                },
                {
                    "name": "ZXzx",
                    "description": "ZXxz",
                    "schema": {
                        "type": "string"
                    },
                    "in": "cookie",
                    "required": true
                }
            ]
        },
        "/abcs/{abcId}": {
            "summary": "Path used to manage a single abc.",
            "description": "The REST endpoint/path used to get, update, and delete single instances of an `abc`.  This path contains `GET`, `PUT`, and `DELETE` operations used to perform the get, update, and delete tasks, respectively.",
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/abc"
                                }
                            }
                        },
                        "description": "Successful response - returns a single `abc`."
                    }
                },
                "operationId": "getabc",
                "summary": "Get a abc",
                "description": "Gets the details of a single instance of a `abc`."
            },
            "put": {
                "requestBody": {
                    "description": "Updated `abc` information.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/abc"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful response."
                    }
                },
                "operationId": "updateabc",
                "summary": "Update a abc",
                "description": "Updates an existing `abc`."
            },
            "delete": {
                "responses": {
                    "204": {
                        "description": "Successful response."
                    }
                },
                "operationId": "deleteabc",
                "summary": "Delete a abc",
                "description": "Deletes an existing `abc`."
            },
            "parameters": [
                {
                    "name": "abcId",
                    "description": "A unique identifier for a `abc`.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        }
    },
    "components": {
        "schemas": {
            "dataSetList": {
                "type": "object",
                "properties": {
                    "total": {
                        "type": "integer"
                    },
                    "apis": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "apiKey": {
                                    "description": "To be used as a dataset parameter value",
                                    "type": "string"
                                },
                                "apiVersionNumber": {
                                    "description": "To be used as a version parameter value",
                                    "type": "string"
                                },
                                "apiUrl": {
                                    "format": "uriref",
                                    "description": "The URL describing the dataset's fields",
                                    "type": "string"
                                },
                                "apiDocumentationUrl": {
                                    "format": "uriref",
                                    "description": "A URL to the API console for each API",
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "example": {
                    "total": 2,
                    "apis": [
                        {
                            "apiKey": "some text",
                            "apiVersionNumber": "some text",
                            "apiUrl": "",
                            "apiDocumentationUrl": ""
                        },
                        {
                            "apiKey": "some text",
                            "apiVersionNumber": "some text",
                            "apiUrl": "",
                            "apiDocumentationUrl": ""
                        }
                    ]
                }
            },
            "abc": {
                "description": "",
                "required": [
                    "xzx"
                ],
                "type": "object",
                "properties": {
                    "xzx": {
                        "description": "",
                        "enum": [
                            "zxX",
                            "fbcxbc"
                        ],
                        "type": "string"
                    }
                },
                "example": {
                    "xzx": "fbcxbc"
                },
                "x-SFSDF": {
                    "xzx": "fbcxbc"
                }
            }
        },
        "responses": {
            "fasf": {
                "headers": {
                    "cZxzcz": {
                        "schema": {
                            "type": "array",
                            "items": {

                            }
                        },
                        "examples": {
                            "xczxczx": {
                                "value": [
                                    {

                                    },
                                    {

                                    }
                                ]
                            }
                        }
                    },
                    "zcxczc": {
                        "schema": {
                            "enum": [
                                "xczczxcz",
                                "zczxczx"
                            ],
                            "type": "string"
                        },
                        "examples": {
                            "cxzczxc": {
                                "value": "\"zczxczx\""
                            }
                        }
                    },
                    "cxzxcx": {

                    }
                },
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "enum": [
                                "xzczxcz",
                                "cgxvxc"
                            ],
                            "type": "string"
                        },
                        "examples": {
                            "zxczxc": {
                                "value": "zxczczxxcz"
                            }
                        }
                    }
                },
                "description": ""
            }
        }
    },
    "tags": [
        {
            "name": "metadata",
            "description": "Find out about the data sets"
        },
        {
            "name": "search",
            "description": "Search a data set"
        }
    ]
}